+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - error object to `docsig.messages`
 - option to check methods belonging to protected classes
 
+### Removed
+- hint constants in `docsig.messages`
+
 [0.40.0](https://github.com/jshwi/docsig/releases/tag/v0.40.0) - 2023-11-25
 ------------------------------------------------------------------------
 ### Added